Pull all leaves of parser into own files
